requirements-dev.txt
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.1
defusedxml==0.6.0
Django==2.2.7
django-braces==1.14.0
django-cors-headers==3.1.1
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.9.1
djangorestframework-datatables==0.4.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
Faker==0.9.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==1.2.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
MarkupPy==1.14
MarkupSafe==1.1.1
mixer==6.1.3
more-itertools==8.0.2
oauthlib==3.1.0
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==3.0.2
packaging==19.2
Pillow==6.2.1
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.5
pytest==5.3.4
pytest-django==3.7.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-json-logger==0.1.11
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rest-social-auth==2.2.0
simplejson==3.17.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
structlog==19.2.0
tablib==0.14.0
text-unidecode==1.2
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
xlrd==1.2.0
XlsxWriter==1.2.6
xlwt==1.3.0
zipp==0.6.0
